pyproject.toml
[project]
name = "hcmai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clip",
    "colorama>=0.4.6",
    "configparser>=7.2.0",
    "einops>=0.8.1",
    "fairscale>=0.4.13",
    "ffmpeg-python>=0.2.0",
    "huggingface-hub>=0.33.0",
    "imagehash>=4.3.2",
    "ipywidgets>=8.1.7",
    "jiwer>=3.1.0",
    "joblib>=1.5.1",
    "jsonlines>=4.0.0",
    "jupyter>=1.1.1",
    "librosa>=0.11.0",
    "lvis>=0.5.3",
    "mmcv==2.2.0",
    "mmengine==0.10.5",
    "moviepy>=2.2.1",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "peft>=0.16.0",
    "pillow>=11.2.1",
    "pycocoevalcap>=1.2",
    "pydub>=0.25.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "scipy>=1.15.3",
    "sentencepiece>=0.2.0",
    "shapely>=2.1.1",
    "tensorflow>=2.18.1",
    "terminaltables>=3.1.10",
    "textgrid>=1.6.1",
    "timm==0.4.12",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "tqdm>=4.67.1",
    "transformers==4.37.2",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
